loadtests/requirements.txt
locustio==0.8.1
requests==2.19.1
